+++ network/src/archive/mod.rs
@@ -5,5 +5,3 @@ pub mod pack;
 pub mod process;
 pub mod orchestrator;
 pub mod sync;
-
-pub use queue::SegmentJob;
